+++ packages/server-services/sql/views/completed_course_exam_count_view.sql
@@ -1,7 +1,7 @@
 SELECT
     cicv.user_id,
     cicv.course_id,
-    COUNT(*) completed_exam_count
+    COUNT(*)::int completed_exam_count
 FROM public.course_item_completion_view cicv
 
 WHERE cicv.exam_version_id IS NOT NULL
