+++ packages/server-services/sql/views/completed_course_item_count_view.sql
@@ -1,7 +1,7 @@
 SELECT
     cicv.user_id,
     cicv.course_id,
-    COUNT(*) completed_course_item_count
+    COUNT(*)::int completed_course_item_count
 FROM public.course_item_completion_view cicv
 
 GROUP BY cicv.user_id, cicv.course_id
