+++ packages/server-services/sql/views/user_learning_page_stats_view.sql
@@ -68,7 +68,7 @@ SELECT
 	-- videos to be repeated count
 	(
 		SELECT 
-			COALESCE(SUM(is_recommended_for_practise::int), 0)
+			COALESCE(SUM(is_recommended_for_practise::int), 0)::int
 		FROM public.user_practise_recommendation_view uprv
 		WHERE uprv.user_id = u.id
 	) videos_to_be_repeated_count,
